Moving some code around to make the file manager operate independently of all other stuff. Added some test code to run and debug a lean version of the application under Cygwin.

+++ software/filesystem/fat_fs.cc
@@ -935,7 +935,7 @@ Directory *FATFS::dir_open(FileInfo *f)
 {
 //    f->print_info();
     FATDIR *fd = new FATDIR(this);
-    Directory *d = new Directory(this, (uint32_t)fd);
+    Directory *d = new Directory(this, fd);
     
     FRESULT res = fd->open(f);
     if(res == FR_OK)
